#180 Refactor: add some method and remove redundant method
